signTxPoolRegistration: update for stax
